+++ include/util.h
@@ -23,14 +23,6 @@ const char *osqp_version(void);
  * Utility Functions  *
  **********************/
 
-/**
- * Set default settings from constants.h file
- * assumes settings already allocated in memory
- * @param settings settings structure
- */
-void set_default_settings(OSQPSettings * settings);
-
-
 #ifndef EMBEDDED
 
 /**
@@ -80,6 +72,7 @@ void print_polish(OSQPWorkspace * work);
  * @param polish is polish enabled?
  */
 void print_footer(OSQPInfo * info, c_int polish);
+
 
 #endif