+++ src/osqp.c
@@ -21,6 +21,45 @@
  * Main API Functions *
  **********************/
 
+void osqp_set_default_settings(OSQPSettings * settings) {
+    settings->scaling = SCALING; /* heuristic problem scaling */
+
+#if EMBEDDED != 1
+    settings->adaptive_rho = ADAPTIVE_RHO;
+    settings->adaptive_rho_interval = ADAPTIVE_RHO_INTERVAL;
+    settings->adaptive_rho_tolerance = (c_float) ADAPTIVE_RHO_TOLERANCE;
+#ifdef PROFILING
+    settings->adaptive_rho_fraction = (c_float) ADAPTIVE_RHO_FRACTION;
+#endif
+#endif
+
+    settings->rho = (c_float) RHO; /* ADMM step */
+    settings->sigma = (c_float) SIGMA; /* ADMM step */
+    settings->max_iter = MAX_ITER; /* maximum iterations to take */
+    settings->eps_abs = (c_float) EPS_ABS;         /* absolute convergence tolerance */
+    settings->eps_rel = (c_float) EPS_REL;         /* relative convergence tolerance */
+    settings->eps_prim_inf = (c_float) EPS_PRIM_INF;         /* primal infeasibility tolerance */
+    settings->eps_dual_inf = (c_float) EPS_DUAL_INF;         /* dual infeasibility tolerance */
+    settings->alpha = (c_float) ALPHA;     /* relaxation parameter */
+    settings->linsys_solver = LINSYS_SOLVER;     /* relaxation parameter */
+
+#ifndef EMBEDDED
+    settings->delta = DELTA;    /* regularization parameter for polish */
+    settings->polish = POLISH;     /* ADMM solution polish: 1 */
+    settings->polish_refine_iter = POLISH_REFINE_ITER; /* iterative refinement
+                                                    steps in polish */
+    settings->verbose = VERBOSE;     /* print output */
+#endif
+
+    settings->scaled_termination = SCALED_TERMINATION;     /* Evaluate scaled termination criteria*/
+    settings->check_termination = CHECK_TERMINATION;     /* Interval for evaluating termination criteria */
+    settings->warm_start = WARM_START;     /* warm starting */
+
+#ifdef PROFILING
+    settings->time_limit = TIME_LIMIT;
+#endif
+}
+
 
 #ifndef EMBEDDED
 
@@ -31,7 +70,7 @@ OSQPWorkspace * osqp_setup(const OSQPData * data, OSQPSettings *settings){
 	// Validate data
 	if (validate_data(data)){
 #ifdef PRINTING
-		c_print("ERROR: Data validation returned failure!\n");
+		c_eprint("Data validation returned failure");
 #endif
 		return OSQP_NULL;
 	}
@@ -39,7 +78,7 @@ OSQPWorkspace * osqp_setup(const OSQPData * data, OSQPSettings *settings){
 	// Validate settings
 	if (validate_settings(settings)){
 #ifdef PRINTING
-		c_print("ERROR: Settings validation returned failure!\n");
+		c_eprint("Settings validation returned failure");
 #endif
 		return OSQP_NULL;
 	}
@@ -48,7 +87,7 @@ OSQPWorkspace * osqp_setup(const OSQPData * data, OSQPSettings *settings){
 	work = c_calloc(1, sizeof(OSQPWorkspace));
 	if (!work){
 #ifdef PRINTING
-		c_print("ERROR: allocating work failure!\n");
+		c_eprint("allocating work failure");
 #endif
 		return OSQP_NULL;
 	}
@@ -131,7 +170,7 @@ OSQPWorkspace * osqp_setup(const OSQPData * data, OSQPSettings *settings){
 	// Load linear system solver
 	if ( load_linsys_solver(work->settings->linsys_solver) ) {
 #ifdef PRINTING
-		c_print("%s linear system solver not available.\nTried to obtain it from shared library.\n",
+		c_eprint("%s linear system solver not available.\nTried to obtain it from shared library",
 				LINSYS_SOLVER_NAME[work->settings->linsys_solver]);
 #endif
 		osqp_cleanup(work);
@@ -144,7 +183,7 @@ OSQPWorkspace * osqp_setup(const OSQPData * data, OSQPSettings *settings){
 			work->settings->linsys_solver, 0);
 	if (!work->linsys_solver) {
 #ifdef PRINTING
-		c_print("ERROR: Linear systems solver initialization failure!\n");
+		c_eprint("Linear systems solver initialization failure");
 #endif
 		osqp_cleanup(work);
 		return OSQP_NULL;
@@ -218,26 +257,35 @@ OSQPWorkspace * osqp_setup(const OSQPData * data, OSQPSettings *settings){
 
 
 c_int osqp_solve(OSQPWorkspace * work){
-	c_int exitflag = 0;
+	c_int exitflag;
 	c_int iter;
 	c_int compute_cost_function;  // Boolean whether to compute the cost function
 	// in the loop
-	c_int can_check_termination = 0;  // Boolean whether to check termination
+	c_int can_check_termination;  // Boolean whether to check termination
 #ifdef PROFILING
 	c_float temp_run_time;  // Temporary variable to store current run time
 #endif
-
 #ifdef PRINTING
 	c_int can_print; // Boolean whether you can print
+#endif
+
+	// Initialize variables
+	exitflag = 0;
+	can_check_termination = 0;
+#ifdef PRINTING
+	can_print = work->settings->verbose;
+#endif
+#ifdef PRINTING
 	compute_cost_function = work->settings->verbose; // Compute cost function only if verbose is on
 #else
 	compute_cost_function = 0; // Never compute cost function during the iterations if no printing enabled
 #endif
 
+	
 	// Check if workspace has been initialized
 	if (!work){
 #ifdef PRINTING
-		c_print("ERROR: Workspace not initialized!\n");
+		c_eprint("Workspace not initialized");
 #endif
 		return -1;
 	}
@@ -245,6 +293,7 @@ c_int osqp_solve(OSQPWorkspace * work){
 #ifdef PROFILING
 	tic(work->timer); // Start timer
 #endif
+
 
 #ifdef PRINTING
 	if (work->settings->verbose){
@@ -363,9 +412,16 @@ c_int osqp_solve(OSQPWorkspace * work){
 		if (work->settings->adaptive_rho && !work->settings->adaptive_rho_interval){
 			// Check time 
 			if (toc(work->timer) > work->settings->adaptive_rho_fraction * work->info->setup_time){
-					// Enough time has passed. We round the number of iterations to the
-					// closest multiple of check_termination
-					work->settings->adaptive_rho_interval = (c_int)c_roundmultiple(iter, work->settings->check_termination);
+					// Enough time has passed. We now get the number of iterations between the updates.
+					if(work->settings->check_termination){	
+						// If check_termination is enabled, we round the number of iterations between 
+						// rho updates to the closest multiple of check_termination
+						work->settings->adaptive_rho_interval = (c_int)c_roundmultiple(iter, work->settings->check_termination);
+					} else {
+						// If check_termintion is disabled, we round the number of iterations between
+						// updates to the closest multiple of the default check_termination interval.
+						work->settings->adaptive_rho_interval = (c_int)c_roundmultiple(iter, CHECK_TERMINATION);
+					}
 					// Make sure the interval is not 0 and at least check_termination times
 					work->settings->adaptive_rho_interval = c_max(work->settings->adaptive_rho_interval, work->settings->check_termination);
 					
@@ -400,7 +456,7 @@ c_int osqp_solve(OSQPWorkspace * work){
 			// Actually update rho
 			if(adapt_rho(work)){
 #ifdef PRINTING
-				c_print("ERROR: Failed rho update!\n");
+				c_eprint("Failed rho update");
 #endif  // PRINTING
 				exitflag = 1;
 				goto exit;
@@ -551,7 +607,11 @@ c_int osqp_cleanup(OSQPWorkspace * work){
 		}
 
 		// Free linear system solver structure
-		work->linsys_solver->free(work->linsys_solver);
+		if (work->linsys_solver){
+			if(work->linsys_solver->free){
+				work->linsys_solver->free(work->linsys_solver);
+			}
+		}
 
 		// Unload linear system solver
 		exitflag = unload_linsys_solver(work->settings->linsys_solver);
@@ -646,7 +706,7 @@ c_int osqp_update_bounds(OSQPWorkspace * work, c_float * l_new, c_float * u_new)
 	for (i=0; i<work->data->m; i++) {
 		if (l_new[i] > u_new[i]) {
 #ifdef PRINTING
-			c_print("lower bound must be lower than or equal to upper bound\n");
+			c_eprint("lower bound must be lower than or equal to upper bound");
 #endif
 			return 1;
 		}
@@ -689,7 +749,7 @@ c_int osqp_update_lower_bound(OSQPWorkspace * work, c_float * l_new) {
 	for (i=0; i<work->data->m; i++) {
 		if (work->data->l[i] > work->data->u[i]) {
 #ifdef PRINTING
-			c_print("upper bound must be greater than or equal to lower bound\n");
+			c_eprint("upper bound must be greater than or equal to lower bound");
 #endif
 			return 1;
 		}
@@ -723,7 +783,7 @@ c_int osqp_update_upper_bound(OSQPWorkspace * work, c_float * u_new) {
 	for (i=0; i<work->data->m; i++) {
 		if (work->data->u[i] < work->data->l[i]) {
 #ifdef PRINTING
-			c_print("lower bound must be lower than or equal to upper bound\n");
+			c_eprint("lower bound must be lower than or equal to upper bound");
 #endif
 			return 1;
 		}
@@ -839,15 +899,16 @@ c_int osqp_update_P(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx, 
 		// nonzeros in P
 		if (P_new_n > nnzP){
 #ifdef PRINTING
-			c_print("Error in P update: new number of elements (%i) greater than elements in P (%i)!\n", (int)P_new_n, (int)nnzP);
+			c_eprint("new number of elements (%i) greater than elements in P (%i)", (int)P_new_n, (int)nnzP);
 #endif
 			return 1;
 		}
 	}
 
-	// Unscale data
-	unscale_data(work);
-
+	if (work->settings->scaling) {
+            // Unscale data
+	    unscale_data(work);
+	}
 
 	// Update P elements
 	if (Px_new_idx){ // Change only Px_new_idx
@@ -861,9 +922,11 @@ c_int osqp_update_P(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx, 
 			work->data->P->x[i] = Px_new[i];
 		}
 	}
-
-	// Scale data
-	scale_data(work);
+	
+	if (work->settings->scaling) {
+            // Scale data
+	    scale_data(work);
+	}
 
 	// Update linear system structure with new data
 	exitflag = work->linsys_solver->update_matrices(work->linsys_solver, work->data->P, work->data->A, work->settings);
@@ -873,7 +936,7 @@ c_int osqp_update_P(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx, 
 
 #ifdef PRINTING
 	if (exitflag < 0) {
-		c_print("Error in P update: new KKT matrix is not quasidefinite!");
+		c_eprint("new KKT matrix is not quasidefinite");
 	}
 #endif
 
@@ -908,14 +971,16 @@ c_int osqp_update_A(OSQPWorkspace * work, c_float * Ax_new, c_int * Ax_new_idx, 
 		// nonzeros in A
 		if (A_new_n > nnzA){
 #ifdef PRINTING
-			c_print("Error in A update: new number of elements (%i) greater than elements in A (%i)!\n", (int)A_new_n, (int)nnzA);
+			c_eprint("new number of elements (%i) greater than elements in A (%i)", (int)A_new_n, (int)nnzA);
 #endif
 			return 1;
 		}
 	}
 
-	// Unscale data
-	unscale_data(work);
+	if (work->settings->scaling) {
+  	    // Unscale data
+	    unscale_data(work);
+	}
 
 	// Update A elements
 	if (Ax_new_idx){ // Change only Ax_new_idx
@@ -929,8 +994,10 @@ c_int osqp_update_A(OSQPWorkspace * work, c_float * Ax_new, c_int * Ax_new_idx, 
 		}
 	}
 
-	// Scale data
-	scale_data(work);
+	if (work->settings->scaling) {
+	    // Scale data
+	    scale_data(work);
+	}
 
 	// Update linear system structure with new data
 	exitflag = work->linsys_solver->update_matrices(work->linsys_solver, work->data->P, work->data->A, work->settings);
@@ -940,7 +1007,7 @@ c_int osqp_update_A(OSQPWorkspace * work, c_float * Ax_new, c_int * Ax_new_idx, 
 
 #ifdef PRINTING
 	if (exitflag < 0) {
-		c_print("Error in A update: new KKT matrix is not quasidefinite!");
+		c_eprint("new KKT matrix is not quasidefinite");
 	}
 #endif
 
@@ -987,7 +1054,7 @@ c_int osqp_update_P_A(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx
 		// nonzeros in P
 		if (P_new_n > nnzP){
 #ifdef PRINTING
-			c_print("Error in P update: new number of elements (%i) greater than elements in P (%i)!\n", (int)P_new_n, (int)nnzP);
+			c_eprint("new number of elements (%i) greater than elements in P (%i)", (int)P_new_n, (int)nnzP);
 #endif
 			return 1;
 		}
@@ -999,15 +1066,16 @@ c_int osqp_update_P_A(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx
 		// nonzeros in A
 		if (A_new_n > nnzA){
 #ifdef PRINTING
-			c_print("Error in A update: new number of elements (%i) greater than elements in A (%i)!\n", (int)A_new_n, (int)nnzA);
+			c_eprint("new number of elements (%i) greater than elements in A (%i)", (int)A_new_n, (int)nnzA);
 #endif
 			return 2;
 		}
 	}
 
-
-	// Unscale data
-	unscale_data(work);
+	if (work->settings->scaling) {
+            // Unscale data
+	    unscale_data(work);
+	}
 
 	// Update P elements
 	if (Px_new_idx){ // Change only Px_new_idx
@@ -1034,9 +1102,10 @@ c_int osqp_update_P_A(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx
 		}
 	}
 
-
-	// Scale data
-	scale_data(work);
+	if (work->settings->scaling) {
+            // Scale data
+	    scale_data(work);
+	}
 
 	// Update linear system structure with new data
 	exitflag = work->linsys_solver->update_matrices(work->linsys_solver, work->data->P, work->data->A, work->settings);
@@ -1046,7 +1115,7 @@ c_int osqp_update_P_A(OSQPWorkspace * work, c_float * Px_new, c_int * Px_new_idx
 
 #ifdef PRINTING
 	if (exitflag < 0) {
-		c_print("Error in P and A update: new KKT matrix is not quasidefinite!");
+		c_eprint("new KKT matrix is not quasidefinite");
 	}
 #endif
 
@@ -1061,7 +1130,7 @@ c_int osqp_update_rho(OSQPWorkspace * work, c_float rho_new){
 	// Check value of rho
 	if (rho_new <= 0) {
 #ifdef PRINTING
-		c_print("rho must be positive\n");
+		c_eprint("rho must be positive");
 #endif
 		return 1;
 	}
@@ -1101,7 +1170,7 @@ c_int osqp_update_max_iter(OSQPWorkspace * work, c_int max_iter_new) {
 	// Check that max_iter is positive
 	if (max_iter_new <= 0) {
 #ifdef PRINTING
-		c_print("max_iter must be positive\n");
+		c_eprint("max_iter must be positive");
 #endif
 		return 1;
 	}
@@ -1114,9 +1183,9 @@ c_int osqp_update_max_iter(OSQPWorkspace * work, c_int max_iter_new) {
 
 c_int osqp_update_eps_abs(OSQPWorkspace * work, c_float eps_abs_new) {
 	// Check that eps_abs is positive
-	if (eps_abs_new <= 0.) {
+	if (eps_abs_new < 0.) {
 #ifdef PRINTING
-		c_print("eps_abs must be positive\n");
+		c_eprint("eps_abs must be nonnegative");
 #endif
 		return 1;
 	}
@@ -1128,9 +1197,9 @@ c_int osqp_update_eps_abs(OSQPWorkspace * work, c_float eps_abs_new) {
 
 c_int osqp_update_eps_rel(OSQPWorkspace * work, c_float eps_rel_new) {
 	// Check that eps_rel is positive
-	if (eps_rel_new <= 0.) {
+	if (eps_rel_new < 0.) {
 #ifdef PRINTING
-		c_print("eps_rel must be positive\n");
+		c_eprint("eps_rel must be nonnegative");
 #endif
 		return 1;
 	}
@@ -1143,9 +1212,9 @@ c_int osqp_update_eps_rel(OSQPWorkspace * work, c_float eps_rel_new) {
 c_int osqp_update_eps_prim_inf(OSQPWorkspace * work, c_float eps_prim_inf_new){
 
 	// Check that eps_prim_inf is positive
-	if (eps_prim_inf_new <= 0.) {
+	if (eps_prim_inf_new < 0.) {
 #ifdef PRINTING
-		c_print("eps_prim_inf must be positive\n");
+		c_eprint("eps_prim_inf must be nonnegative");
 #endif
 		return 1;
 	}
@@ -1160,9 +1229,9 @@ c_int osqp_update_eps_prim_inf(OSQPWorkspace * work, c_float eps_prim_inf_new){
 c_int osqp_update_eps_dual_inf(OSQPWorkspace * work, c_float eps_dual_inf_new){
 
 	// Check that eps_dual_inf is positive
-	if (eps_dual_inf_new <= 0.) {
+	if (eps_dual_inf_new < 0.) {
 #ifdef PRINTING
-		c_print("eps_dual_inf must be positive\n");
+		c_eprint("eps_dual_inf must be nonnegative");
 #endif
 		return 1;
 	}
@@ -1178,7 +1247,7 @@ c_int osqp_update_alpha(OSQPWorkspace * work, c_float alpha_new) {
 	// Check that alpha is between 0 and 2
 	if (alpha_new <= 0. || alpha_new >= 2.) {
 #ifdef PRINTING
-		c_print("alpha must be between 0 and 2\n");
+		c_eprint("alpha must be between 0 and 2");
 #endif
 		return 1;
 	}
@@ -1193,7 +1262,7 @@ c_int osqp_update_warm_start(OSQPWorkspace * work, c_int warm_start_new) {
 	// Check that warm_start is either 0 or 1
 	if (warm_start_new != 0 && warm_start_new != 1) {
 #ifdef PRINTING
-		c_print("warm_start should be either 0 or 1\n");
+		c_eprint("warm_start should be either 0 or 1");
 #endif
 		return 1;
 	}
@@ -1208,7 +1277,7 @@ c_int osqp_update_scaled_termination(OSQPWorkspace * work, c_int scaled_terminat
 	// Check that scaled_termination is either 0 or 1
 	if (scaled_termination_new != 0 && scaled_termination_new != 1) {
 #ifdef PRINTING
-		c_print("scaled_termination should be either 0 or 1\n");
+		c_eprint("scaled_termination should be either 0 or 1");
 #endif
 		return 1;
 	}
@@ -1222,7 +1291,7 @@ c_int osqp_update_check_termination(OSQPWorkspace * work, c_int check_terminatio
 	// Check that check_termination is nonnegative
 	if (check_termination_new < 0) {
 #ifdef PRINTING
-		c_print("check_termination should be nonnegative\n");
+		c_eprint("check_termination should be nonnegative");
 #endif
 		return 1;
 	}
@@ -1239,7 +1308,7 @@ c_int osqp_update_delta(OSQPWorkspace * work, c_float delta_new) {
 	// Check that delta is positive
 	if (delta_new <= 0.) {
 #ifdef PRINTING
-		c_print("delta must be positive\n");
+		c_eprint("delta must be positive");
 #endif
 		return 1;
 	}
@@ -1253,7 +1322,7 @@ c_int osqp_update_polish(OSQPWorkspace * work, c_int polish_new) {
 	// Check that polish is either 0 or 1
 	if (polish_new != 0 && polish_new != 1) {
 #ifdef PRINTING
-		c_print("polish should be either 0 or 1\n");
+		c_eprint("polish should be either 0 or 1");
 #endif
 		return 1;
 	}
@@ -1273,7 +1342,7 @@ c_int osqp_update_polish_refine_iter(OSQPWorkspace * work, c_int polish_refine_i
 	// Check that polish_refine_iter is nonnegative
 	if (polish_refine_iter_new < 0) {
 #ifdef PRINTING
-		c_print("polish_refine_iter must be nonnegative\n");
+		c_eprint("polish_refine_iter must be nonnegative");
 #endif
 		return 1;
 	}
@@ -1288,7 +1357,7 @@ c_int osqp_update_verbose(OSQPWorkspace * work, c_int verbose_new) {
 	// Check that verbose is either 0 or 1
 	if (verbose_new != 0 && verbose_new != 1) {
 #ifdef PRINTING
-		c_print("verbose should be either 0 or 1\n");
+		c_eprint("verbose should be either 0 or 1");
 #endif
 		return 1;
 	}